Some little GUI changes...
Starting VolumeGroups to group Output Channels into Stereo/Tripple/Quad/5.1/etc... outputmodes.
Include a mainobject for a group, a widget for the master properties and a widget for every channel. Placing an audio source into the different channels could be done via complex squareroots which all have the abs of 1 from the middle... more toughts to follow...